+++ design-system/ds-lib/src/index.tsx
@@ -1,3 +1,7 @@
-// Nothing exported here, import directly from individual entry-points
+// This file exists for module resolution purposes.
 
-export default undefined;
+// Not exporting anything from the root entry point
+// Please use individual entry points in order to always
+// obtain minimum kbs.
+
+export default {};