+++ design-system/skeleton/examples/constellation/skeleton-basic.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import Skeleton from '../../src';
+
+export default () => <Skeleton width="200px" height="16px" testId="skeleton" />;