+++ design-system/skeleton/examples/constellation/skeleton-shimmering.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import Skeleton from '../../src';
+
+export default () => (
+  <Skeleton width="200px" height="16px" isShimmering testId="skeleton" />
+);