+++ drag-and-drop/docs/src/index.ts
@@ -1,4 +1,4 @@
 /**
  * This file exists for module resolution purposes.
  */
-export default undefined;
+export default {};