+++ drag-and-drop/hitbox/src/index.ts
@@ -1,5 +1,7 @@
+// This file exists for module resolution purposes.
+
 // Not exporting anything from the root entry point
 // Please use individual entry points in order to always
-// obtain minimum kbs
+// obtain minimum kbs.
 
-export default undefined;
+export default {};