+++ editor/editor-core/src/plugins/expand/__tests__/integration/node-selection/expand-collapse.ts
@@ -1,5 +1,4 @@
-import { runBlockNodeSelectionTestSuite } from '../../../../../__tests__/integration/selection/_blockNodeSelectionTestSuite';
-
+import { runBlockNodeSelectionTestSuite } from '@atlaskit/editor-test-helpers/integration/selection';
 import { default as WebDriverPage } from '@atlaskit/webdriver-runner/wd-wrapper';
 
 // helper function to collapse all expands being tested
@@ -32,14 +31,11 @@ runBlockNodeSelectionTestSuite({
   skipTests: {
     'Click and drag from the end to start of the document to select [block-node] when [block-node] is the first node':
       [
-        /* Drag and drop selection with Expand (collpased) is not stable enough to allow us to trust on integration tests */
+        /* Drag and drop selection with Expand (collapsed) is not stable enough to allow us to trust on integration tests */
         'firefox',
         'chrome',
         'safari',
       ],
-    'Click and drag from start to end of document and select [block-node]': [
-      'firefox',
-    ],
     'Extend selection up one line to select [block-node] with shift + arrow up':
       ['firefox'],
     'Extend selection right two characters to select [block-node] from line above with shift + arrow right':