+++ editor/editor-core/src/plugins/extension/__tests__/integration/block-node-selection.ts
@@ -1,4 +1,5 @@
-import { runBlockNodeSelectionTestSuite } from '../../../../__tests__/integration/selection/_blockNodeSelectionTestSuite';
+import { runBlockNodeSelectionTestSuite } from '@atlaskit/editor-test-helpers/integration/selection';
+
 runBlockNodeSelectionTestSuite({
   nodeName: 'bodiedExtension',
   selector: '.bodiedExtensionView-content-wrap',
@@ -24,8 +25,5 @@ runBlockNodeSelectionTestSuite({
       ['firefox'],
     'Click and drag from the end to start of the document to select [block-node] when [block-node] is the first node':
       ['firefox'],
-    'Click and drag from start to end of document and select [block-node]': [
-      'firefox',
-    ],
   },
 });