+++ editor/editor-core/src/plugins/list/__tests__/integration/block-node-selection/bullet-list.ts
@@ -1,4 +1,5 @@
-import { runBlockNodeSelectionTestSuite } from '../../../../../__tests__/integration/selection/_blockNodeSelectionTestSuite';
+import { runBlockNodeSelectionTestSuite } from '@atlaskit/editor-test-helpers/integration/selection';
+
 runBlockNodeSelectionTestSuite({
   nodeName: 'bulletList',
   editorOptions: { allowLists: true },